refactor(View): Input, Output 분리
