+++ src/main/java/RacingMain.java
@@ -3,25 +3,26 @@ import domain.Count;
 import domain.RaceCars;
 import domain.RacingGame;
 import util.ExceptionHandler;
-import view.View;
+import view.InputView;
+import view.OutputView;
 
 import java.util.List;
 
 public class RacingMain {
 
     public static void main(String[] args) {
-        final View view = new View();
-
-        final RaceCars raceCars = ExceptionHandler.run(view::readCars);
-        final Count count = ExceptionHandler.run(view::readCount);
+        final InputView inputView = new InputView();
+        final RaceCars raceCars = ExceptionHandler.run(inputView::readCars);
+        final Count count = ExceptionHandler.run(inputView::readCount);
 
         final RacingGame racingGame = new RacingGame(raceCars, count);
         racingGame.race();
 
-        view.printResultNotice();
-        view.printRace(racingGame);
+        final OutputView outputView = new OutputView();
+        outputView.printResultNotice();
+        outputView.printRace(racingGame);
 
         final List<Car> winners = racingGame.judgeWinners();
-        view.printWinners(winners);
+        outputView.printWinners(winners);
     }
 }
